requirements.txt
about-time==4.2.1
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alive-progress==3.1.5
annotated-types==0.6.0
anyio==3.7.1
anykeystore==0.2
apex==0.9.10.dev0
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
blis==0.4.1
boto3==1.28.30
botocore==1.31.30
bravado==11.0.3
bravado-core==6.1.0
cachetools==5.3.1
catalogue==1.0.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
colorama==0.4.6
contourpy==1.1.0
cryptacular==1.6.2
cryptography==41.0.3
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
cymem==2.0.7
dataclasses==0.6
datasets==2.14.5
defusedxml==0.7.1
dill==0.3.7
distro==1.9.0
easydict==1.11
ekphrasis==0.5.1
en-core-web-sm==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.12.4
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gensim==3.8.1
geomloss==0.2.6
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
GPUtil==1.4.0
grapheme==0.6.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.17.2
hupper==1.12
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
isoduration==20.11.0
jaraco.classes==3.3.0
jeepney==0.8.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==24.2.0
kiwisolver==1.4.5
knockknock==0.1.7
lime==0.2.0.1
lit==16.0.6
llvmlite==0.41.0
ltp==4.2.13
ltp-core==0.1.4
ltp-extension==0.1.11
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.5.0
matrix-client==0.4.0
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.6
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
neptune==1.6.0
neptune-client==0.4.107
networkx==3.1
nltk==3.8.1
numba==0.58.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==1.10.0
opencv-python==4.9.0.80
openprompt==1.0.1
openpyxl==3.1.2
packaging==23.1
pandas==1.3.0
PasteDeploy==3.0.1
pbkdf2==1.3
perspective==1.0.3
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
plaster==1.1.2
plaster-pastedeploy==1.0.1
plotly==5.17.0
POT==0.9.1
premailer==3.10.0
preshed==3.0.8
protobuf==4.24.3
psutil==5.9.5
py3nvml==0.2.7
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
pynndescent==0.5.10
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyramid==2.0.1
pyramid-mailer==0.15.1
python-dateutil==2.8.2
python-telegram-bot==20.4
python3-openid==3.2.0
pytz==2023.3
PyWavelets==1.1.1
PyYAML==6.0.1
ray==2.6.3
referencing==0.30.2
regex==2023.8.8
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rouge==1.0.0
rpds-py==0.10.3
rsa==4.9
s3transfer==0.6.2
sacremoses==0.0.53
safetensors==0.3.3
scikit-image==0.17.2
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.13.0
SecretStorage==3.3.3
sentencepiece==0.1.96
setuptools-scm==7.1.0
simplejson==3.19.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
spacy==2.3.2
SQLAlchemy==2.0.20
srsly==1.0.7
swagger-spec-validator==3.0.3
sympy==1.12
tbb==2021.10.0
tenacity==8.2.3
tensorboardX==2.6.2.2
termcolor==2.3.0
thinc==7.4.1
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.12.1
tomli==2.0.1
torch==1.8.1+cu111
toxigen==1.1.0
tqdm==4.66.1
transaction==3.1.0
transformers==4.22.0
translationstring==1.4
triton==2.0.0
twilio==8.5.0
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
umap-learn==0.5.4
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.5
velruse==1.1.1
venusian==3.0.0
waiting==1.4.1
wasabi==0.10.1
wcwidth==0.2.6
webcolors==1.13
WebOb==1.8.7
websocket-client==1.6.1
WTForms==3.0.1
wtforms-recaptcha==0.3.2
xmltodict==0.13.0
xxhash==3.4.1
yacs==0.1.8
yagmail==0.15.293
yarl==1.9.2
zipp==3.16.2
zope.deprecation==5.0
zope.interface==6.0
zope.sqlalchemy==3.0
